pyproject.toml
[project]
name = "itk-project"
version = "0.1.0"
description = "Wallet FastAPI Project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.135.2",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = [
    "ruff>=0.3.0",
    "pyright>=1.1.350",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "testcontainers[postgresql]>=3.7.0",
    "httpx>=0.27.0",
]

[tool.ruff]
line-length = 88
target-version = "py312"
src = ["app"]
lint.select = ["E", "F", "I", "N", "UP", "B"] 
lint.ignore = ["B008"]
# E, F - ошибки и стиль, I - сортировка импортов, N - нейминг, UP - апгрейд синтаксиса, B - баги

[tool.pyright]
include = ["app"]
exclude = ["**/node_modules", "**/__pycache__", ".venv", "migrations"]
typeCheckingMode = "basic" # Для начала "basic", чтобы не сойти с ума от строгости
pythonVersion = "3.12"
venvPath = "."       # Искать папку с окружением в текущем корне
venv = ".venv"       # Имя самой папки окружения
